refactor: split interactive cell state and controls
